requirements.txt
adrf==0.1.7 
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

aiohttp==3.10.3
aioredis==2.0.1
asgiref==3.8.1 
celery==5.4.0 
django-celery-beat==2.7.0 
channels==4.1.0 

pymemcache==4.0.0
redis==5.0.8 
django-redis==5.4.0 
hiredis==3.0.0 

asent==0.8.3  
beautifulsoup4==4.12.3 
country-named-entity-recognition==0.4
groq==0.9.0
pycountry==22.1.10
scikit-learn==1.5.2 
spacy==3.8.0
textstat==0.7.4 
vaderSentiment==3.3.2 

pandas==2.2.2
numpy==1.26.4

psycopg==3.2.1 
psycopg-pool==3.2.1

certifi==2024.7.4 
h2==4.1.0 
pyOpenSSL==24.2.1
 
Twisted==24.7.0 

daphne==4.1.2 
Hypercorn==0.17.3 
gunicorn==23.0.0
uvicorn==0.30.6

pipreqs==0.5.0
python-dotenv==1.0.1

mjml-python==1.3.3 